agents/fraud-trends/agent.py
#!/usr/bin/env python3
"""
Fraud Trends Investigator Agent

This agent researches insurance fraud trends using AI and web search capabilities.
It executes a 6-step workflow to generate comprehensive fraud analysis reports.

Usage:
    python agent.py --topic "Auto Insurance Fraud" --regions "US,CA" --time-range "2024-2025"
"""

import argparse
import os
import sys
from typing import Dict, Any
from dotenv import load_dotenv


def parse_arguments() -> argparse.Namespace:
    """
    Parse command-line arguments for the Fraud Trends agent.

    Returns:
        argparse.Namespace: Parsed arguments containing topic, regions, time_range, and optional focus_areas
    """
    parser = argparse.ArgumentParser(
        description="Fraud Trends Investigator Agent - Research insurance fraud trends",
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog="""
Examples:
  python agent.py --topic "Auto Insurance Fraud" --regions "US,CA" --time-range "2024-2025"
  python agent.py --topic "Property Fraud" --regions "FL,TX" --time-range "2024" --focus-areas "climate,claims"
        """
    )

    parser.add_argument(
        "--topic",
        type=str,
        required=True,
        help="The fraud topic to research (e.g., 'Auto Insurance Fraud', 'Synthetic Identity Fraud')"
    )

    parser.add_argument(
        "--regions",
        type=str,
        required=True,
        help="Comma-separated list of regions to focus on (e.g., 'US,CA,UK')"
    )

    parser.add_argument(
        "--time-range",
        type=str,
        required=True,
        dest="time_range",
        help="Time range for the research (e.g., '2024-2025', '2024')"
    )

    parser.add_argument(
        "--focus-areas",
        type=str,
        required=False,
        dest="focus_areas",
        help="Optional comma-separated list of focus areas (e.g., 'detection,prevention,technology')"
    )

    return parser.parse_args()


def validate_environment() -> bool:
    """
    Validate required environment variables are present.

    Returns:
        bool: True if all required variables are present, False otherwise
    """
    required_vars = ["ANTHROPIC_API_KEY", "TAVILY_API_KEY"]
    missing_vars = []

    for var in required_vars:
        if not os.getenv(var):
            missing_vars.append(var)

    if missing_vars:
        print("❌ ERROR: Missing required environment variables:")
        print()
        for var in missing_vars:
            print(f"   - {var}")
        print()
        print("Please create a .env file with the following variables:")
        print()
        for var in missing_vars:
            print(f"   {var}=your_{var.lower()}_here")
        print()
        print("Refer to .env.example for a template.")
        print("See README.md for instructions on getting API keys.")
        print()
        return False

    return True


def main() -> int:
    """
    Main entry point for the Fraud Trends Investigator agent.

    Returns:
        int: Exit code (0 for success, non-zero for failure)
    """
    print("=" * 80)
    print("FRAUD TRENDS INVESTIGATOR AGENT")
    print("=" * 80)
    print()

    # Load environment variables from .env file
    load_dotenv()

    # Validate environment variables
    if not validate_environment():
        return 1

    print("✅ Environment variables validated")
    print()

    # Parse command-line arguments
    args = parse_arguments()

    # Convert comma-separated strings to lists
    regions = [r.strip() for r in args.regions.split(",")]
    focus_areas = None
    if args.focus_areas:
        focus_areas = [f.strip() for f in args.focus_areas.split(",")]

    # Display input parameters
    print("📋 Input Parameters:")
    print(f"   Topic: {args.topic}")
    print(f"   Regions: {', '.join(regions)}")
    print(f"   Time Range: {args.time_range}")
    if focus_areas:
        print(f"   Focus Areas: {', '.join(focus_areas)}")
    print()

    # TODO: 6-step workflow implementation will be added in Epic 3

    print("✅ Agent skeleton successfully executed")
    print("⚠️  Full workflow implementation coming in Epic 3")
    print()

    return 0


if __name__ == "__main__":
    sys.exit(main())
